Added ID based on millis current system time and reorder module structure
