+++ src/event/hash.rs
